belal_shop/sundorban/requirements.txt
asgiref==3.4.1
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.4
Django==3.2.12
django-ckeditor==6.2.0
django-countries==7.3.2
django-crispy-forms==1.13.0
django-js-asset==2.0.0
django-quill-editor==0.1.21
django-tinymce==3.4.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.4.0
pylint==2.7.4
pytz==2021.3
requests==2.22.0
six==1.13.0
sqlparse==0.4.2
sslcommerz-python==0.0.7
toml==0.10.2
typed-ast==1.4.0
typing_extensions==4.1.1
urllib3==1.25.7
wrapt==1.11.2
